Added missing copyright headers; renamed package murmor to murmur
