+++ internal/murmur/murmur_test.go
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package murmor
+package murmur
 
 import (
 	"testing"
